🚧 Add type alias in app.tsx
