+++ 03/first-ts-app/src/App.tsx
@@ -3,8 +3,22 @@ import CountryList from './components/CountryList'
 
 type Props = {}
 
-const App = (props: Props) => {
+export type CountryType = {
+  no: number;
+  country: string;
+  visited: boolean;
+};
+
+
+const App = () => {
   let msg = 'React typescript'
+  let list: CountryType[] = [
+    { no: 1, country: "일본", visited: true },
+    { no: 2, country: "미국", visited: false },
+    { no: 3, country: "영국", visited: true },
+    { no: 4, country: "스위스", visited: false },
+  ];
+
   const addResult = (x: number, y: number) => {
     return (
       <div className='card card-body bg-light mb-3'>
@@ -18,7 +32,7 @@ const App = (props: Props) => {
       <h2>Hello {msg}!</h2>
       <hr className='dash-style' />
       {addResult(4, 3)}
-      <CountryList/>
+      <CountryList countries={list} />
     </div>
   )
 }
